Move headers to src (#10)
