+++ test/test_evt_pin_listener/test_evt_pin_listener.cpp
@@ -25,6 +25,8 @@ void setUp(void)
     targetWithoutDebounce.enable();
 }
 
+void tearDown(void) {}
+
 void test_does_not_trigger_when_disabled(void)
 {
     target.disable();
