+++ test/test_evt_time_listener/test_evt_time_listener.cpp
@@ -26,6 +26,8 @@ void setUp(void)
     target.enable();
 }
 
+void tearDown(void) {}
+
 void test_does_not_trigger_when_disabled(void)
 {
     target.disable();
